logic moved to custom hook
